test case for the exception-after-bulk bug
After having parsed a bulk response, for example the response to GET,
if an unhandled exception happens in the callback the parser is left
in an inconsistent state. Further calls to redis-client inside a global
exception handler will then result in the parser hanging and the
callback not being called.