fix for the exception-after-bulk bug
See the previous commit for the bug description.

Resetting the state machine before calling the callback ensures that
if an exception handler calls redis-client the parser will be in a
clean state.

This is useful for example for using an uncaughtException handler to
clean up when a series of tests fail. The clean up might include calls
to redis, for example to FLUSHDB.